Fixed bugs with game and getPlayerChoice
game function was not taking input for 5 rounds, fixed this by calling both choices function in it.
getPlayerChoice was going in infinite loop when encounter wrong inputs, fixed this by getting input inside the loop.

Also, added player and computer score and checks the score at the end of the game and on comparison declare the wins.